Version 2.13
Added new "AT" command which finds the alternate file and opens in a new tab. Similarly to the "A" and "AS" commands, if the buffer is already open it simply switches to that buffer in the corresponding tab/window. This script will only work with VIM7 from this version forward.